Hide the subword_indices method
Introduce the enum WordIndex to represent either a single word index or
multiple subword indices. Both cases are now handled by the Vocab::idx
method.

There is a small overhead to returning an Option<WordIndex> over an
Option<usize>, however it does not introduce a heap allocation for known
words.