Do not read the embedding file twice in ReadText::read_text
Plain-text embedding files do not have the embedding matrix shape on
the first line. We worked around by doing a first pass over the file
to get the shape. However, this is likely to be quite a bit more
expensive than amortized O(1) appends to Vec. So, instead read all the
embedding components into a Vec; do some sanity checks on the Vec's
final size; move the Vec into an ndarray Array2.