Specialize NGrams to string slices
NGram::new now takes a string slice, rather than a Vec of a generic
type. The resulting iterator returns the n-grams as string slices.

In SubwordIndices::subword_indices, we used to compute the hash of an
n-gram over Vec<char>. The implementation of Hash for Vec first adds
the length of the Vec to the hasher, followed by the individual
element. However, we want to avoid collecting the characters in a Vec
or iterating over the characters twice in order to get the length.
For this reason, NGrams now couples string slices with their length in
characters using the auxiliary StrWithCharLen type.